+++ src/squint/set.js
@@ -24,7 +24,7 @@ export function intersection(...xs) {
   switch (xs.length) {
     case 0: return null;
     case 1: return xs[0];
-    case 2: return xs[0].length > xs[1].length ? 
+    case 2: return xs[0].size > xs[1].size ?
       _intersection2(xs[0], xs[1]) :
       _intersection2(xs[1], xs[0]);
     default: return _bubble_max_key((x) => 0 - x.size, xs).reduce(_intersection2);
@@ -62,8 +62,8 @@ export function union(...xs) {
   switch (xs.length) {
     case 0: return null;
     case 1: return xs[0];
-    case 2: return xs[0].length > xs[1].length ? 
-      _union2(xs[0], xs[1]) : 
+    case 2: return xs[0].size > xs[1].size ?
+      _union2(xs[0], xs[1]) :
       _union2(xs[1], xs[0]);
     default: return _bubble_max_key((x) => x.size, xs).reduce(_union2);
   }